+++ packages/protocol-kit/src/managers/fallbackHandlerManager.ts
@@ -1,6 +1,5 @@
 import { EthAdapter, GnosisSafeContract } from '@safe-global/safe-core-sdk-types'
-import { SAFE_FEATURES, hasSafeFeature } from '../utils'
-import { isZeroAddress, sameString } from '../utils'
+import { SAFE_FEATURES, hasSafeFeature, isZeroAddress, sameString } from '../utils'
 import { ZERO_ADDRESS } from '../utils/constants'
 
 class FallbackHandlerManager {
