+++ packages/protocol-kit/src/managers/ownerManager.ts
@@ -1,5 +1,5 @@
 import { EthAdapter, GnosisSafeContract } from '@safe-global/safe-core-sdk-types'
-import { isRestrictedAddress, sameString } from '../utils'
+import { isRestrictedAddress, sameString } from '../utils/address'
 import { SENTINEL_ADDRESS } from '../utils/constants'
 
 class OwnerManager {
