+++ packages/protocol-kit/src/utils/transactions/utils.ts
@@ -7,6 +7,8 @@ import {
   GnosisSafeContract,
   MetaTransactionData,
   OperationType,
+  SafeMultisigTransactionResponse,
+  SafeTransaction,
   SafeTransactionData,
   SafeTransactionDataPartial
 } from '@safe-global/safe-core-sdk-types'
@@ -113,4 +115,16 @@ export function decodeMultiSendData(encodedData: string): MetaTransactionData[] 
   }
 
   return txs
+}
+
+export function isMetaTransactionArray(
+  safeTransactions: SafeTransactionDataPartial | MetaTransactionData[]
+): safeTransactions is MetaTransactionData[] {
+  return (safeTransactions as MetaTransactionData[])?.length !== undefined
+}
+
+export function isSafeMultisigTransactionResponse(
+  safeTransaction: SafeTransaction | SafeMultisigTransactionResponse
+): safeTransaction is SafeMultisigTransactionResponse {
+  return (safeTransaction as SafeMultisigTransactionResponse).isExecuted !== undefined
 }
